Implement enumerate_by_*_full at the top level
Don't push this into the modules; the membership information may be
stored in one module, and the detailed user/group information in
a different one.  Each module could implement the recursive lookup,
but that's usually just a waste.

(Well, doing this in the modules does help in one case: If an user's
primary group is stored as GID, the module can call lu_group_lookup_id()
on the ID directly; doing this at the top level means a lu_group_lookup_id()
call to get the name, and then a lu_group_lookup_name() call to get the
details.  So far there seems no reason to worry about speed that much.)