files.c: Init char *name to NULL
Merges:
https://pagure.io/libuser/issue/27

This is mostly to silence coverity warnings. "enum lu_entity_type" has
three values and several places in the code follow logic as:

char *name;
if ent->type == user:
        name = foo()
if ent->type == group
        name = bar()
g_assert(name != NULL)

it shouldn't be possible for ent->type to be anything else but in the
odd case it is, initializing name to NULL will ensure that name will be
still NULL after the code falls through the conditions and at least the
behaviour is defined.