Correct popt memory handling
In the code of libusers we use popt library but we release popt context
too early. In older versions, popt leaked memory, so it worked anyway.

With this patch poptFeeContext() call is moved to the end of main()
function.

The patch also unifies the way utilities terminate in case of an error.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=2125904